+++ scout/constants/indexes.py
@@ -84,6 +84,15 @@ INDEXES = {
             ('hpo_number', ASCENDING)],
             name="number"),
     ],
+    'event': [
+        IndexModel([
+            ('category', ASCENDING),
+            ('verb', ASCENDING),
+            ],
+            name="category_verb",
+            background=True,
+            ),
+    ],
     'transcript': [
         IndexModel([
             ('build', ASCENDING),
@@ -108,6 +117,10 @@ INDEXES = {
             background=True,
             ),
     ],
-    
+    'case': [
+        IndexModel([
+            ('synopsis', TEXT)],
+            default_language='english',
+            name="synopsis_text"),
+    ],
 }
-