+++ scout/server/utils.py
@@ -50,8 +50,7 @@ def institute_and_case(store, institute_id, case_name=None):
             if not case_name or not any(inst_id in case_obj['collaborators'] for inst_id in
                                         current_user.institutes):
                 # you don't have access!!
-                flash("You don't have acccess to: {}".format(institute_obj['display_name']),
-                      'danger')
+                flash("You don't have acccess to: {}".format(institute_id),'danger')
                 return abort(403)
 
     # you have access!